improved update PFN logic

+++ Raythena/actors/esworker.py
@@ -191,10 +191,10 @@ class ESWorker:
         ranges = json.loads(ranges_update['eventRanges'][0])
         ranges = EventRangeUpdate.build_from_dict(self.job.get_id(), ranges)
         for range_update in ranges[self.job.get_id()]:
-            cfile = range_update.get("path", "")
-            dst = os.path.join(harvester_endpoint, cfile)
-            range_update["path"] = dst
-            if os.path.isfile(cfile):
+            cfile = range_update.get("path", None)
+            if cfile and os.path.isfile(cfile):
+                dst = os.path.join(harvester_endpoint, os.path.basename(cfile) if os.path.isabs(cfile) else cfile)
+                range_update["path"] = dst
                 self.logging_actor.debug.remote(self.id, f"Moving {cfile} to {dst}")
                 shutil.move(cfile, dst)
         return ranges
